Clean RewardItem を別ファイル RewardItem.kt に切り出し
RewardListScreen.kt が572行に肥大化していたため、RewardItem Composable
と関連 Preview を RewardItem.kt に分離。visibility は同パッケージ内の
RewardList から呼ばれるため internal を維持。
